Refactored API calls to custom hooks to be able to call for repeated API calls from various components for refresh.
